kira::backend::cpal

Struct CpalBackend

pub struct CpalBackend { /* private fields */ }

A backend that uses cpal to connect a Renderer to the operating system’s audio driver.

Implementations

impl CpalBackend

pub fn pop_cpu_usage(&mut self) -> Option<f32>

Returns the oldest reported CPU usage in the queue.

The formula for the CPU usage is time elapsed / time allotted, where

time elapsed is the amount of time it took to fill the audio buffer requested by the OS

time allotted is the maximum amount of time Kira could take to process audio and still finish in time to avoid audio stuttering (num frames / sample rate)

pub fn pop_error(&mut self) -> Option<StreamError>

Returns the oldest available stream error in the queue.

pub fn num_stream_errors_discarded(&self) -> Option<u64>

Returns the number of unhandled stream errors discarded because of overcrowding. This increases either:

When cpal produces errors faster than they are polled by kira, or

When kira produces errors faster than they are polled by Self::pop_error.