kira::backend::cpal
Struct CpalBackendSettings

pub struct CpalBackendSettings {
    pub device: Option<Device>,
    pub buffer_size: BufferSize,
}

Settings for the cpal backend.

Fields
device: Option<Device>

The output audio device to use. If None, the default output device will be used.

buffer_size: BufferSize

The buffer size used by the device. If it is set to BufferSize::Default, the default buffer size for the device will be used. Note that the default buffer size might be surprisingly large, leading to latency issues. If a lower latency is desired, consider using BufferSize::Fixed in accordance with the cpal::SupportedBufferSize range provided by the cpal::SupportedStreamConfig API.