kira::clock
Module clock_info

Contains types for reporting information about clocks.

You’ll only need these types if you’re creating implementations of the Sound, Effect, or Modulator traits. If you want to access information about clocks from gameplay code, use a ClockHandle.

Structs

ClockInfo

Information about the current state of a clock.

ClockInfoProvider

Provides information about any clock that currently exists.

MockClockInfoProviderBuilder

Builds a ClockInfoProvider that provides fake clock info.

Enums

WhenToStart

When something should start given the current state of the clocks.