kira

Module effect

Modifies audio signals.

Any type that implements EffectBuilder can be added to a mixer track by using TrackBuilder::add_effect. Kira comes with a number of commonly used effects.

If needed, you can create custom effects by implementing the EffectBuilder and Effect traits.

Modules

compressor

Reduces (or increases) the dynamic range of audio.

delay

Adds echoes to a sound.

distortion

Makes a sound harsher and noisier.

eq_filter

Adjusts the volume of a frequency range of a sound.

filter

Removes frequencies from a sound.

panning_control

Adjusts the panning of audio.

reverb

Adds reverberations to a sound. Useful for simulating room tones.

volume_control

Adjusts the volume of audio.

Traits

Effect

Receives input audio from a mixer track and outputs modified audio.

EffectBuilder

Configures an effect.