kira

Module info

Types for providing info about resources to trait implementors.

You’ll only need this if you’re implementing one of Kira’s traits, like Sound or Effect.

Structs

ClockInfo

Information about the current state of a clock.

Info

Provides info about resources on the audio thread.

ListenerInfo

Information about a listener.

MockInfoBuilder

Generates a fake Info with arbitrary data. Useful for writing unit tests.

Enums

WhenToStart

When something should start given the current state of the clocks.