kira::info
Struct MockInfoBuilder

pub struct MockInfoBuilder { /* private fields */ }

Generates a fake Info with arbitrary data. Useful for writing unit tests.

Implementations

impl MockInfoBuilder

pub fn new() -> Self

Creates a new MockInfoBuilder.

pub fn add_clock(&mut self, ticking: bool, ticks: u64, fraction: f64) -> ClockId

Adds a fake clock with the given ticking state and time. Returns a fake ClockId.

pub fn add_modulator(&mut self, value: f64) -> ModulatorId

Adds a fake modulator outputting the given value. Returns a fake ModulatorId.

pub fn add_listener(&mut self, listener_info: ListenerInfo) -> ListenerId

Adds a fake listener at the given position and orientation. Returns a fake ListenerId.

pub fn build(self) -> Info<'static>

Consumes the MockInfoProvider and returns a fake Info.