Trait kira::manager::backend::Backend

pub trait Backend: Sized {
    type Settings;
    type Error;

    // Required methods
    fn setup(settings: Self::Settings) -> Result<(Self, u32), Self::Error>;
    fn start(&mut self, renderer: Renderer) -> Result<(), Self::Error>;
}

Connects a Renderer to a lower level audio API.

Required Associated Types

type Settings

Settings for this backend.

type Error

Errors that can occur when using this backend.

Required Methods

fn setup(settings: Self::Settings) -> Result<(Self, u32), Self::Error>

Starts the backend and returns itself and the initial sample rate.

fn start(&mut self, renderer: Renderer) -> Result<(), Self::Error>

Sends the renderer to the backend to start audio playback.

Object Safety

This trait is not object safe.