Struct kira::manager::AudioManager

pub struct AudioManager<B: Backend = DefaultBackend> { /* private fields */ }

Controls audio from gameplay code.

Implementations

impl<B: Backend> AudioManager<B>

pub fn new(settings: AudioManagerSettings<B>) -> Result<Self, B::Error>

Creates a new AudioManager.

Examples

Create an AudioManager using the DefaultBackend and the default settings:

use kira::manager::{AudioManager, AudioManagerSettings, backend::DefaultBackend};

let audio_manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings::default())?;

Create an AudioManager with a reverb effect on the main mixer track:

use kira::{
	manager::{AudioManager, AudioManagerSettings, backend::DefaultBackend},
	track::TrackBuilder,
	effect::reverb::ReverbBuilder,
};

let audio_manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings {
	main_track_builder: TrackBuilder::new().with_effect(ReverbBuilder::new()),
	..Default::default()
})?;

pub fn play<D: SoundData>( &mut self, sound_data: D, ) -> Result<D::Handle, PlaySoundError<D::Error>>

Plays a sound.

Examples

use kira::sound::static_sound::{StaticSoundData, StaticSoundSettings};

let sound_data = StaticSoundData::from_file("sound.ogg")?;
manager.play(sound_data)?;

pub fn add_sub_track( &mut self, builder: TrackBuilder, ) -> Result<TrackHandle, ResourceLimitReached>

Creates a mixer sub-track.

pub fn add_clock( &mut self, speed: impl Into<Value<ClockSpeed>>, ) -> Result<ClockHandle, ResourceLimitReached>

Creates a clock.

Examples

let clock = manager.add_clock(ClockSpeed::TicksPerMinute(120.0))?;

pub fn add_spatial_scene( &mut self, settings: SpatialSceneSettings, ) -> Result<SpatialSceneHandle, ResourceLimitReached>

Creates a spatial scene.

pub fn add_modulator<Builder: ModulatorBuilder>( &mut self, builder: Builder, ) -> Result<Builder::Handle, ResourceLimitReached>

Creates a modulator.

Examples

use kira::modulator::lfo::LfoBuilder;

let modulator = manager.add_modulator(LfoBuilder::new())?;

pub fn main_track(&mut self) -> &mut TrackHandle

Returns a handle to the main mixer track.

Examples

Use the main track handle to adjust the volume of all audio:

use kira::tween::Tween;

manager.main_track().set_volume(0.5, Tween::default());

pub fn sound_capacity(&self) -> u16

Returns the number of sounds that can be loaded at a time.

pub fn sub_track_capacity(&self) -> u16

Returns the number of mixer sub-tracks that can exist at a time.

pub fn clock_capacity(&self) -> u16

Returns the number of clocks that can exist at a time.

pub fn spatial_scene_capacity(&self) -> u16

Returns the number of spatial scenes that can exist at a time.

pub fn modulator_capacity(&self) -> u16

Returns the number of modulators that can exist at a time.

pub fn num_sounds(&self) -> u16

Returns the number of sounds that are currently loaded.

pub fn num_sub_tracks(&self) -> u16

Returns the number of mixer sub-tracks that currently exist.

pub fn num_clocks(&self) -> u16

Returns the number of clocks that currently exist.

pub fn num_spatial_scenes(&self) -> u16

Returns the number of spatial scenes that currently exist.

pub fn num_modulators(&self) -> u16

Returns the number of modulators that currently exist.

pub fn backend_mut(&mut self) -> &mut B

Returns a mutable reference to this manager’s backend.