kira::modulator::lfo

Struct LfoBuilder

pub struct LfoBuilder {
    pub waveform: Waveform,
    pub frequency: Value<f64>,
    pub amplitude: Value<f64>,
    pub offset: Value<f64>,
    pub starting_phase: f64,
}

Configures an LFO modulator.

Fields
waveform: Waveform

The oscillation pattern.

frequency: Value<f64>

How quickly the value oscillates.

amplitude: Value<f64>

How much the value oscillates.

An amplitude of 2.0 means the modulator will reach a maximum value of 2.0 and a minimum value of -2.0.

offset: Value<f64>

The constant value the modulator is offset by.

An LFO with an offset of 1.0 and an amplitude of 0.5 will reach a maximum value of 1.5 and a minimum value of 0.5.

starting_phase: f64

The phase the LFO should start at (in radians).

This determines when in the oscillation the modulator will start.

Implementations

impl LfoBuilder

pub fn new() -> Self

Creates a new LfoBuilder with the default settings.

pub fn waveform(self, waveform: Waveform) -> Self

Sets the oscillation pattern.

pub fn frequency(self, frequency: impl Into<Value<f64>>) -> Self

Sets how quickly the value oscillates.

pub fn amplitude(self, amplitude: impl Into<Value<f64>>) -> Self

Sets how much the value oscillates.

An amplitude of 2.0 means the modulator will reach a maximum value of 2.0 and a minimum value of -2.0.

pub fn offset(self, offset: impl Into<Value<f64>>) -> Self

Sets a constant value that the modulator is offset by.

An LFO with an offset of 1.0 and an amplitude of 0.5 will reach a maximum value of 1.5 and a minimum value of 0.5.

pub fn starting_phase(self, starting_phase: f64) -> Self

Sets the phase the LFO should start at (in radians).

This determines when in the oscillation the modulator will start.