Struct kira::sound::static_sound::StaticSoundSettings

pub struct StaticSoundSettings {
    pub start_time: StartTime,
    pub start_position: PlaybackPosition,
    pub loop_region: Option<Region>,
    pub reverse: bool,
    pub volume: Value<Volume>,
    pub playback_rate: Value<PlaybackRate>,
    pub panning: Value<f64>,
    pub output_destination: OutputDestination,
    pub fade_in_tween: Option<Tween>,
}

Settings for a static sound.

Fields
start_time: StartTime

When the sound should start playing.

start_position: PlaybackPosition

Where in the sound playback should start.

loop_region: Option<Region>

The portion of the sound that should be looped.

reverse: bool

Whether the sound should be played in reverse.

volume: Value<Volume>

The volume of the sound.

playback_rate: Value<PlaybackRate>

The playback rate of the sound.

Changing the playback rate will change both the speed and the pitch of the sound.

panning: Value<f64>

The panning of the sound, where 0 is hard left and 1 is hard right.

output_destination: OutputDestination

The destination that this sound should be routed to.

fade_in_tween: Option<Tween>

An optional fade-in from silence.

Implementations

impl StaticSoundSettings

pub fn new() -> Self

Creates a new StaticSoundSettings with the default settings.

pub fn start_time(self, start_time: impl Into<StartTime>) -> Self

Sets when the sound should start playing.

pub fn start_position(self, start_position: impl Into<PlaybackPosition>) -> Self

Sets where in the sound playback should start.

pub fn reverse(self, reverse: bool) -> Self

Sets whether the sound should be played in reverse.

pub fn loop_region(self, loop_region: impl IntoOptionalRegion) -> Self

Sets the portion of the sound that should be looped.

pub fn volume(self, volume: impl Into<Value<Volume>>) -> Self

Sets the volume of the sound.

pub fn playback_rate( self, playback_rate: impl Into<Value<PlaybackRate>>, ) -> Self

Sets the playback rate of the sound.

Changing the playback rate will change both the speed and the pitch of the sound.

pub fn panning(self, panning: impl Into<Value<f64>>) -> Self

Sets the panning of the sound, where 0 is hard left and 1 is hard right.

pub fn output_destination( self, output_destination: impl Into<OutputDestination>, ) -> Self

Sets the destination that this sound should be routed to.

pub fn fade_in_tween(self, fade_in_tween: impl Into<Option<Tween>>) -> Self

Sets the tween used to fade in the sound from silence.