kira
Struct AudioManager

pub struct AudioManager<B: Backend = DefaultBackend> { /* private fields */ }

Controls audio from gameplay code.

Implementations

impl<B: Backend> AudioManager<B>

pub fn new(settings: AudioManagerSettings<B>) -> Result<Self, B::Error>

Creates a new AudioManager.

Examples

Create an AudioManager using the DefaultBackend and the default settings:

use kira::{AudioManager, AudioManagerSettings, DefaultBackend};

let audio_manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings::default())?;

Create an AudioManager with a reverb effect on the main mixer track:

use kira::{
	AudioManager, AudioManagerSettings, DefaultBackend,
	track::MainTrackBuilder,
	effect::reverb::ReverbBuilder,
};

let audio_manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings {
	main_track_builder: MainTrackBuilder::new().with_effect(ReverbBuilder::new()),
	..Default::default()
})?;

pub fn play<D: SoundData>( &mut self, sound_data: D, ) -> Result<D::Handle, PlaySoundError<D::Error>>

Plays a sound.

Examples

use kira::sound::static_sound::{StaticSoundData, StaticSoundSettings};

let sound_data = StaticSoundData::from_file("sound.ogg")?;
manager.play(sound_data)?;

pub fn add_sub_track( &mut self, builder: TrackBuilder, ) -> Result<TrackHandle, ResourceLimitReached>

Creates a mixer sub-track.

pub fn add_spatial_sub_track( &mut self, listener: impl Into<ListenerId>, position: impl Into<Value<Vector3<f32>>>, builder: SpatialTrackBuilder, ) -> Result<SpatialTrackHandle, ResourceLimitReached>

Adds a spatial mixer sub-track.

pub fn add_send_track( &mut self, builder: SendTrackBuilder, ) -> Result<SendTrackHandle, ResourceLimitReached>

Creates a mixer send track.

pub fn add_clock( &mut self, speed: impl Into<Value<ClockSpeed>>, ) -> Result<ClockHandle, ResourceLimitReached>

Creates a clock.

Examples

use kira::clock::ClockSpeed;

let clock = manager.add_clock(ClockSpeed::TicksPerMinute(120.0))?;

pub fn add_modulator<Builder: ModulatorBuilder>( &mut self, builder: Builder, ) -> Result<Builder::Handle, ResourceLimitReached>

Creates a modulator.

Examples

use kira::modulator::lfo::LfoBuilder;

let modulator = manager.add_modulator(LfoBuilder::new())?;

pub fn add_listener( &mut self, position: impl Into<Value<Vector3<f32>>>, orientation: impl Into<Value<Quaternion<f32>>>, ) -> Result<ListenerHandle, ResourceLimitReached>

Creates a listener.

Examples

// This example uses `glam`, but you can use any math library that has interoperability
// with `mint`.
let listener = manager.add_listener(glam::vec3(0.0, 0.0, 0.0), glam::Quat::IDENTITY)?;

pub fn main_track(&mut self) -> &mut MainTrackHandle

Returns a handle to the main mixer track.

Examples

Use the main track handle to adjust the volume of all audio:

use kira::Tween;

manager.main_track().set_volume(-6.0, Tween::default());

pub fn sub_track_capacity(&self) -> usize

Returns the number of mixer sub-tracks that can exist at a time.

pub fn send_track_capacity(&self) -> usize

Returns the number of mixer send tracks that can exist at a time.

pub fn clock_capacity(&self) -> usize

Returns the number of clocks that can exist at a time.

pub fn modulator_capacity(&self) -> usize

Returns the number of modulators that can exist at a time.

pub fn num_sub_tracks(&self) -> usize

Returns the number of mixer sub-tracks that currently exist.

pub fn num_send_tracks(&self) -> usize

Returns the number of mixer send tracks that currently exist.

pub fn num_clocks(&self) -> usize

Returns the number of clocks that currently exist.

pub fn num_modulators(&self) -> usize

Returns the number of modulators that currently exist.

pub fn backend_mut(&mut self) -> &mut B

Returns a mutable reference to this manager’s backend.