Module kira::tween

Smooth interpolation between values.

Structs

ModulatorMapping

A transformation from a modulator’s value to a parameter value.

Parameter

Manages and updates a value that can be smoothly transitioned and linked to modulators.

Tween

Describes a smooth transition between values.

Enums

Easing

Curves the motion of a Tween.

Value

A value that a parameter can be linked to.

Traits

Tweenable

A trait for types that can be smoothly interpolated.