Struct kira::manager::AudioManager

pub struct AudioManager<CustomEvent: Copy + Send + 'static = ()> { /* fields omitted */ }

Plays and manages audio.

The AudioManager is responsible for all communication between the gameplay thread and the audio thread.

Implementations

impl<CustomEvent: Copy + Send + 'static + Debug> AudioManager<CustomEvent>

pub fn new(settings: AudioManagerSettings) -> KiraResult<Self>

Creates a new audio manager and starts an audio thread.

pub fn add_sound(&mut self, sound: Sound) -> KiraResult<SoundId>

Sends a sound to the audio thread and returns a handle to the sound.

pub fn remove_sound(&mut self, id: SoundId) -> KiraResult<()>

Removes a sound from the audio thread, allowing its memory to be freed.

pub fn play_sound(
&mut self,
sound_id: SoundId,
settings: InstanceSettings
) -> Result<InstanceId, KiraError>

Plays a sound.

pub fn set_instance_volume(
&mut self,
id: InstanceId,
volume: Value
) -> Result<(), KiraError>

Sets the volume of an instance.

pub fn set_instance_pitch(
&mut self,
id: InstanceId,
pitch: Value
) -> Result<(), KiraError>

Sets the pitch of an instance.

pub fn pause_instance(
&mut self,
instance_id: InstanceId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Pauses a currently playing instance of a sound with an optional fade-out tween.

pub fn resume_instance(
&mut self,
instance_id: InstanceId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Resumes a currently paused instance of a sound with an optional fade-in tween.

pub fn stop_instance(
&mut self,
instance_id: InstanceId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Stops a currently playing instance of a sound with an optional fade-out tween.

Once the instance is stopped, it cannot be restarted.

pub fn pause_instances_of_sound(
&mut self,
sound_id: SoundId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Pauses all currently playing instances of a sound with an optional fade-out tween.

pub fn resume_instances_of_sound(
&mut self,
sound_id: SoundId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Resumes all currently playing instances of a sound with an optional fade-in tween.

pub fn stop_instances_of_sound(
&mut self,
sound_id: SoundId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Stops all currently playing instances of a sound with an optional fade-out tween.

pub fn set_metronome_tempo(&mut self, tempo: Tempo) -> Result<(), KiraError>

Sets the tempo of the metronome.

pub fn start_metronome(&mut self) -> Result<(), KiraError>

Starts or resumes the metronome.

pub fn pause_metronome(&mut self) -> Result<(), KiraError>

Pauses the metronome.

pub fn stop_metronome(&mut self) -> Result<(), KiraError>

Stops and resets the metronome.

pub fn start_sequence(
&mut self,
sequence: Sequence<CustomEvent>
) -> Result<SequenceId, KiraError>

Starts a sequence.

pub fn mute_sequence(&mut self, id: SequenceId) -> Result<(), KiraError>

Mutes a sequence.

When a sequence is muted, it will continue waiting for durations and intervals, but it will not play sounds, emit events, or perform any other actions.

pub fn unmute_sequence(&mut self, id: SequenceId) -> Result<(), KiraError>

Unmutes a sequence.

pub fn pause_sequence(&mut self, id: SequenceId) -> Result<(), KiraError>

Pauses a sequence.

pub fn resume_sequence(&mut self, id: SequenceId) -> Result<(), KiraError>

Resumes a sequence.

pub fn stop_sequence(&mut self, id: SequenceId) -> Result<(), KiraError>

Stops a sequence.

pub fn pause_sequence_and_instances(
&mut self,
id: SequenceId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Pauses a sequence and any instances played by that sequence.

pub fn resume_sequence_and_instances(
&mut self,
id: SequenceId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Resumes a sequence and any instances played by that sequence.

pub fn stop_sequence_and_instances(
&mut self,
id: SequenceId,
fade_tween: Option<Tween>
) -> Result<(), KiraError>

Stops a sequence and any instances played by that sequence.

pub fn add_parameter(&mut self, value: f64) -> KiraResult<ParameterId>

Creates a parameter with the specified starting value.

pub fn remove_parameter(&mut self, id: ParameterId) -> KiraResult<()>

Removes a parameter.

pub fn set_parameter(
&mut self,
id: ParameterId,
value: f64,
tween: Option<Tween>
) -> KiraResult<()>

Sets the value of a parameter with an optional tween to smoothly change the value.

pub fn add_sub_track(
&mut self,
settings: TrackSettings
) -> KiraResult<SubTrackId>

Creates a mixer sub-track.

pub fn remove_sub_track(&mut self, id: SubTrackId) -> KiraResult<()>

Removes a sub-track from the mixer.

pub fn add_effect_to_track<T: Into<TrackIndex> + Copy>(
&mut self,
track_index: T,
effect: Box<dyn Effect>,
settings: EffectSettings
) -> KiraResult<EffectId>

Adds an effect to a track.

pub fn remove_effect(&mut self, effect_id: EffectId) -> KiraResult<()>

Removes an effect from the mixer.

pub fn events(&mut self) -> Vec<Event<CustomEvent>>

Returns a list of all of the new events created by the audio thread (since the last time events was called).

pub fn free_unused_resources(&mut self)

Frees resources that are no longer in use, such as unloaded sounds or finished sequences.