kira::modulator

Trait ModulatorBuilder

pub trait ModulatorBuilder {
    type Handle;

    // Required method
    fn build(self, id: ModulatorId) -> (Box<dyn Modulator>, Self::Handle);
}

Configures a modulator.

Required Associated Types

type Handle

Allows the user to control the modulator from gameplay code.

Required Methods

fn build(self, id: ModulatorId) -> (Box<dyn Modulator>, Self::Handle)

Creates the modulator and a handle to the modulator.