kira::sound::streaming

Struct StreamingSoundData

pub struct StreamingSoundData<Error: Send + 'static> {
    pub settings: StreamingSoundSettings,
    pub slice: Option<(usize, usize)>,
    /* private fields */
}

A streaming sound that is not playing yet.

Fields
settings: StreamingSoundSettings

Settings for the streaming sound.

slice: Option<(usize, usize)>

The portion of the sound this StreamingSoundData represents.

Note that the StreamingSoundData holds the entire piece of audio it was originally given regardless of the value of slice, but StreamingSoundData::num_frames and StreamingSoundData::duration will behave as if this StreamingSoundData only contained the specified portion of audio.

Implementations

impl<Error: Send> StreamingSoundData<Error>

pub fn from_decoder(decoder: impl Decoder<Error = Error> + 'static) -> Self

Creates a StreamingSoundData for a Decoder.

pub fn start_time(self, start_time: impl Into<StartTime>) -> Self

Sets when the sound should start playing.

Examples

Configuring a sound to start 4 ticks after a clock’s current time:

use kira::{
	AudioManager, AudioManagerSettings, DefaultBackend,
	sound::streaming::{StreamingSoundData, StreamingSoundSettings},
	clock::ClockSpeed,
};

let mut manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings::default())?;
let clock_handle = manager.add_clock(ClockSpeed::TicksPerMinute(120.0))?;
let sound = StreamingSoundData::from_file("sound.ogg")?
	.start_time(clock_handle.time() + 4);

pub fn start_position(self, start_position: impl Into<PlaybackPosition>) -> Self

Sets where in the sound playback should start.

pub fn loop_region(self, loop_region: impl IntoOptionalRegion) -> Self

Sets the portion of the sound that should be looped.

Examples

Configure a sound to loop the portion from 3 seconds in to the end:

let sound = StreamingSoundData::from_file("sound.ogg")?.loop_region(3.0..);

Configure a sound to loop the portion from 2 to 4 seconds:

let sound = StreamingSoundData::from_file("sound.ogg")?.loop_region(2.0..4.0);

pub fn volume(self, volume: impl Into<Value<Decibels>>) -> Self

Sets the volume of the sound.

Examples

Set the volume to a fixed value:

let sound = StreamingSoundData::from_file("sound.ogg")?.volume(-6.0);

Link the volume to a modulator:

use kira::{
	AudioManager, AudioManagerSettings, DefaultBackend,
	modulator::tweener::TweenerBuilder,
	sound::streaming::StreamingSoundData,
	Value, Mapping, Easing,
	Decibels,
};

let mut manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings::default())?;
let tweener = manager.add_modulator(TweenerBuilder {
	initial_value: 0.5,
})?;
let sound = StreamingSoundData::from_file("sound.ogg")?.volume(Value::FromModulator {
	id: tweener.id(),
	mapping: Mapping {
		input_range: (0.0, 1.0),
		output_range: (Decibels::SILENCE, Decibels::IDENTITY),
		easing: Easing::Linear,
	},
});

pub fn playback_rate( self, playback_rate: impl Into<Value<PlaybackRate>>, ) -> Self

Sets the playback rate of the sound.

Changing the playback rate will change both the speed and the pitch of the sound.

Examples

Set the playback rate as a factor:

let sound = StreamingSoundData::from_file("sound.ogg")?.playback_rate(0.5);

Set the playback rate as a change in semitones:

use kira::Semitones;
let sound = StreamingSoundData::from_file("sound.ogg")?.playback_rate(Semitones(-2.0));

Link the playback rate to a modulator:

use kira::{
	AudioManager, AudioManagerSettings, DefaultBackend,
	modulator::tweener::TweenerBuilder,
	sound::streaming::StreamingSoundData,
	Value, Easing, Mapping,
	PlaybackRate,
};

let mut manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings::default())?;
let tweener = manager.add_modulator(TweenerBuilder {
	initial_value: 0.5,
})?;
let sound = StreamingSoundData::from_file("sound.ogg")?.playback_rate(Value::FromModulator {
	id: tweener.id(),
	mapping: Mapping {
		input_range: (0.0, 1.0),
		output_range: (PlaybackRate(0.0), PlaybackRate(1.0)),
		easing: Easing::Linear,
	},
});

pub fn panning(self, panning: impl Into<Value<Panning>>) -> Self

Sets the panning of the sound, where 0 is hard left and 1 is hard right.

Examples

Set the panning to a fixed value:

let sound = StreamingSoundData::from_file("sound.ogg")?.panning(-0.5);

Link the panning to a modulator:

use kira::{
	AudioManager, AudioManagerSettings, DefaultBackend,
	modulator::tweener::TweenerBuilder,
	sound::streaming::StreamingSoundData,
	Value, Easing, Mapping,
	Panning,
};

let mut manager = AudioManager::<DefaultBackend>::new(AudioManagerSettings::default())?;
let tweener = manager.add_modulator(TweenerBuilder {
	initial_value: -0.5,
})?;
let sound = StreamingSoundData::from_file("sound.ogg")?.panning(Value::FromModulator {
	id: tweener.id(),
	mapping: Mapping {
		input_range: (-1.0, 1.0),
		output_range: (Panning::LEFT, Panning::RIGHT),
		easing: Easing::Linear,
	},
});

pub fn fade_in_tween(self, fade_in_tween: impl Into<Option<Tween>>) -> Self

Sets the tween used to fade in the instance from silence.

pub fn with_settings(self, settings: StreamingSoundSettings) -> Self

Returns the StreamingSoundData with the specified settings.

pub fn num_frames(&self) -> usize

Returns the number of frames in the StreamingSoundData.

If StreamingSoundData::slice is Some, this will be the number of frames in the slice.

pub fn duration(&self) -> Duration

Returns the duration of the audio.

If StreamingSoundData::slice is Some, this will be the duration of the slice.

pub fn unsliced_duration(&self) -> Duration

Returns the total duration of the audio, regardless of its slice.

pub fn slice(self, region: impl IntoOptionalRegion) -> Self

Sets the portion of the audio this StreamingSoundData represents.

impl StreamingSoundData<FromFileError>

pub fn from_file( path: impl AsRef<Path>, ) -> Result<StreamingSoundData<FromFileError>, FromFileError>

Creates a StreamingSoundData for an audio file.

pub fn from_cursor<T: AsRef<[u8]> + Send + Sync + 'static>( cursor: Cursor<T>, ) -> Result<StreamingSoundData<FromFileError>, FromFileError>

Creates a StreamingSoundData for a cursor wrapping audio file data.

pub fn from_media_source( media_source: impl MediaSource + 'static, ) -> Result<StreamingSoundData<FromFileError>, FromFileError>

Creates a StreamingSoundData for a type that implements Symphonia’s MediaSource trait.